add script for select features
